+++ src/Transformers/IntervalDiscretizer.php
@@ -13,6 +13,8 @@ use Rubix\ML\Exceptions\RuntimeException;
 
 use function Rubix\ML\linspace;
 use function array_slice;
+use function min;
+use function max;
 
 /**
  * Interval Discretizer
@@ -36,6 +38,13 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
     protected int $bins;
 
     /**
+     * Should the bins be equal width?
+     *
+     * @var bool
+     */
+    protected bool $equiWidth;
+
+    /**
      * The bin intervals of the fitted data.
      *
      * @var array[]|null
@@ -44,9 +53,10 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
 
     /**
      * @param int $bins
+     * @param bool $equiWidth
      * @throws \Rubix\ML\Exceptions\InvalidArgumentException
      */
-    public function __construct(int $bins = 5)
+    public function __construct(int $bins = 5, bool $equiWidth = false)
     {
         if ($bins < 3) {
             throw new InvalidArgumentException('Number of bins must be'
@@ -54,6 +64,7 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
         }
 
         $this->bins = $bins;
+        $this->equiWidth = $equiWidth;
     }
 
     /**
@@ -98,9 +109,11 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
     {
         SamplesAreCompatibleWithTransformer::with($dataset, $this)->check();
 
-        $q = linspace(0.0, 1.0, 1 + $this->bins);
+        if (!$this->equiWidth) {
+            $q = linspace(0.0, 1.0, $this->bins + 1);
 
-        $q = array_slice($q, 1, -1);
+            $q = array_slice($q, 1, -1);
+        }
 
         $this->intervals = [];
 
@@ -108,7 +121,16 @@ class IntervalDiscretizer implements Transformer, Stateful, Persistable
             if ($type->isContinuous()) {
                 $values = $dataset->feature($column);
 
-                $edges = Stats::quantiles($values, $q);
+                if (isset($q)) {
+                    $edges = Stats::quantiles($values, $q);
+                } else {
+                    $min = min($values);
+                    $max = max($values);
+
+                    $edges = linspace($min, $max, $this->bins + 1);
+
+                    $edges = array_slice($edges, 1, -1);
+                }
 
                 $edges[] = INF;