+++ tests/Classifiers/LogitBoostTest.php
@@ -1,0 +1,195 @@
+<?php
+
+namespace Rubix\ML\Tests\Classifiers;
+
+use Rubix\ML\Learner;
+use Rubix\ML\Verbose;
+use Rubix\ML\DataType;
+use Rubix\ML\Estimator;
+use Rubix\ML\Persistable;
+use Rubix\ML\Probabilistic;
+use Rubix\ML\RanksFeatures;
+use Rubix\ML\EstimatorType;
+use Rubix\ML\Loggers\BlackHole;
+use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Classifiers\LogitBoost;
+use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\Regressors\RegressionTree;
+use Rubix\ML\CrossValidation\Metrics\FBeta;
+use Rubix\ML\Datasets\Generators\Agglomerate;
+use Rubix\ML\CrossValidation\Metrics\Accuracy;
+use Rubix\ML\Exceptions\RuntimeException;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @group Classifiers
+ * @covers \Rubix\ML\Classifiers\LogitBoost
+ */
+class LogitBoostTest extends TestCase
+{
+    /**
+     * The number of samples in the training set.
+     *
+     * @var int
+     */
+    protected const TRAIN_SIZE = 350;
+
+    /**
+     * The number of samples in the validation set.
+     *
+     * @var int
+     */
+    protected const TEST_SIZE = 20;
+
+    /**
+     * The minimum validation score required to pass the test.
+     *
+     * @var float
+     */
+    protected const MIN_SCORE = 0.9;
+
+    /**
+     * Constant used to see the random number generator.
+     *
+     * @var int
+     */
+    protected const RANDOM_SEED = 0;
+
+    /**
+     * @var \Rubix\ML\Datasets\Generators\Agglomerate
+     */
+    protected $generator;
+
+    /**
+     * @var \Rubix\ML\Classifiers\LogitBoost
+     */
+    protected $estimator;
+
+    /**
+     * @var \Rubix\ML\CrossValidation\Metrics\Accuracy
+     */
+    protected $metric;
+
+    /**
+     * @before
+     */
+    protected function setUp() : void
+    {
+        $this->generator = new Agglomerate([
+            'male' => new Blob([69.2, 195.7, 40.0], [2.0, 6.0, 0.6]),
+            'female' => new Blob([63.7, 168.5, 38.1], [1.6, 5.0, 0.8]),
+        ]);
+
+        $this->estimator = new LogitBoost(new RegressionTree(3), 0.1, 0.5, 1000, 1e-4, 5, 0.1, new FBeta());
+
+        $this->metric = new Accuracy();
+
+        srand(self::RANDOM_SEED);
+    }
+
+    /**
+     * @test
+     */
+    public function build() : void
+    {
+        $this->assertInstanceOf(LogitBoost::class, $this->estimator);
+        $this->assertInstanceOf(Estimator::class, $this->estimator);
+        $this->assertInstanceOf(Learner::class, $this->estimator);
+        $this->assertInstanceOf(Probabilistic::class, $this->estimator);
+        $this->assertInstanceOf(RanksFeatures::class, $this->estimator);
+        $this->assertInstanceOf(Verbose::class, $this->estimator);
+        $this->assertInstanceOf(Persistable::class, $this->estimator);
+    }
+
+    /**
+     * @test
+     */
+    public function type() : void
+    {
+        $this->assertEquals(EstimatorType::classifier(), $this->estimator->type());
+    }
+
+    /**
+     * @test
+     */
+    public function compatibility() : void
+    {
+        $expected = [
+            DataType::categorical(),
+            DataType::continuous(),
+        ];
+
+        $this->assertEquals($expected, $this->estimator->compatibility());
+    }
+
+    /**
+     * @test
+     */
+    public function params() : void
+    {
+        $expected = [
+            'min change' => 0.0001,
+            'window' => 5,
+            'booster' => new RegressionTree(3),
+            'rate' => 0.1,
+            'ratio' => 0.5,
+            'estimators' => 1000,
+            'hold out' => 0.1,
+            'metric' => new FBeta(1),
+        ];
+
+        $this->assertEquals($expected, $this->estimator->params());
+    }
+
+    /**
+     * @test
+     */
+    public function trainPredict() : void
+    {
+        $this->estimator->setLogger(new BlackHole());
+
+        $training = $this->generator->generate(self::TRAIN_SIZE);
+        $testing = $this->generator->generate(self::TEST_SIZE);
+
+        $this->estimator->train($training);
+
+        $this->assertTrue($this->estimator->trained());
+
+        $scores = $this->estimator->losses();
+
+        $this->assertIsArray($scores);
+        $this->assertContainsOnly('float', $scores);
+
+        $losses = $this->estimator->losses();
+
+        $this->assertIsArray($losses);
+        $this->assertContainsOnly('float', $losses);
+
+        $importances = $this->estimator->featureImportances();
+
+        $this->assertIsArray($importances);
+        $this->assertCount(3, $importances);
+        $this->assertContainsOnly('float', $importances);
+
+        $predictions = $this->estimator->predict($testing);
+
+        $score = $this->metric->score($predictions, $testing->labels());
+
+        $this->assertGreaterThanOrEqual(self::MIN_SCORE, $score);
+    }
+
+    /**
+     * @test
+     */
+    public function predictUntrained() : void
+    {
+        $this->expectException(RuntimeException::class);
+
+        $this->estimator->predict(Unlabeled::quick());
+    }
+
+    protected function assertPreConditions() : void
+    {
+        $this->assertFalse($this->estimator->trained());
+    }
+}